+++ host/features/eh_host_feat_wifi/src/eh_host_wifi.c
@@ -23,6 +23,7 @@
 #include "esp_wifi_types.h"
 
 #include "eh_host_wifi.h"
+#include "eh_host_wifi_priv.h"
 
 #include "eh_host_feat_rpc_ext_v2_api_common.h"
 
@@ -537,6 +538,11 @@ static bool s_ap_netif_started;
 bool eh_host_wifi_rx_admitted(bool is_ap)
 {
     return s_rx_admit[is_ap ? 1 : 0];
+}
+
+void eh_host_wifi_admit_rx(bool is_ap, bool admit)
+{
+    s_rx_admit[is_ap ? 1 : 0] = admit;
 }
 
 static void wifi_sta_connected_handler(const void *ctrl_cmd, void *ctx)